refactor: qc-application hub 도메인 패키지 정리
