+++ qc-application/src/main/java/com/eager/questioncloud/application/api/hub/review/implement/QuestionHunReviewUpdater.java
@@ -1,4 +1,4 @@
-package com.eager.questioncloud.application.api.hub;
+package com.eager.questioncloud.application.api.hub.review.implement;
 
 import com.eager.questioncloud.core.domain.review.infrastructure.QuestionReviewRepository;
 import com.eager.questioncloud.core.domain.review.model.QuestionReview;
