+++ qc-application/src/test/java/com/eager/questioncloud/application/api/hub/QuestionHubReviewRegisterTest.java
@@ -2,6 +2,7 @@ package com.eager.questioncloud.application.api.hub;
 
 import static org.assertj.core.api.Assertions.assertThat;
 
+import com.eager.questioncloud.application.api.hub.review.implement.QuestionHubReviewRegister;
 import com.eager.questioncloud.core.domain.question.infrastructure.QuestionEntity;
 import com.eager.questioncloud.core.domain.question.infrastructure.QuestionJpaRepository;
 import com.eager.questioncloud.core.domain.question.model.Question;
